api/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12
paddlepaddle==3.0.0
paddleocr==2.9.1
Pillow==11.0.0
aiofiles==24.1.0
